Cast size_t to unsigned long long in error.h
Since the size of size_t is machine dependent,
it may vary between 4 and 8 bytes.
In some machines size_t is a typedef of an
unsigned long and in others, of a unsigned long long.
So to avoid trouble, allways cast the parameter in the printerr macros
to the highest posible size.

+++ src/Util/error.h
@@ -28,7 +28,7 @@
 
 #define printerr_null_param(func) fprintf(stderr, "ERROR: NULL parameter(s) in function " #func "\n");
 
-#define printerr_out_of_bounds(index, func) fprintf(stderr, "ERROR: Index %lu out of bounds. At function " #func "\n", index);
+#define printerr_out_of_bounds(index, func) fprintf(stderr, "ERROR: Index %llu out of bounds. At function " #func "\n", (unsigned long long) index);
 
 #define printerr_memory_op(func) fprintf(stderr, "ERROR: At memory operation in function " #func "\n");
